Change config to dependency and test with dependency overrides
