+++ laa_court_data_api_app/models/ping.py
@@ -1,6 +1,7 @@
 from typing import Optional
 from pydantic import BaseModel, Field
 
+
 class Ping(BaseModel):
     app_branch: Optional[str] = Field(None, example='test_branch')
     build_date: Optional[str] = Field(None, example='02022022')
